+++ src/hooks/useCloseBurgerMenu.js
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+
+export default function useCloseBurgerMenu(ref, callback) {
+  useEffect(() => {
+    /**
+     * Run callback if clicked on outside of element
+     */
+    function handleClickOutside(event) {
+      if (ref.current && !ref.current.contains(event.target)) {
+        callback();
+      }
+    }
+    // Bind the event listener
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      // Unbind the event listener on clean up
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [ref, callback]);
+}
